# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynaptiFlow \u2014 Closed-Loop Discovery MVP",
    "description": "Assay-metadata schema as the spine of wet\u2194dry communication. Boltz outputs are PREDICTIONS, never assay records. All Boltz access goes through app/boltz_client.py with mock mode, caching, and a credit guard.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/loop/replay": {
      "post": {
        "tags": [
          "loop"
        ],
        "summary": "Loop Replay",
        "operationId": "loop_replay_loop_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReplayRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "List Records",
        "operationId": "list_records_records_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssayRecord"
                  },
                  "title": "Response List Records Records Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "records"
        ],
        "summary": "Post Record",
        "operationId": "post_record_records_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssayRecordIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/{record_id}": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Get Record",
        "operationId": "get_record_records__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssayRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "records"
        ],
        "summary": "Patch Record",
        "operationId": "patch_record_records__record_id__patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/compound/{inchikey}": {
      "get": {
        "tags": [
          "compound"
        ],
        "summary": "Get Compound",
        "operationId": "get_compound_compound__inchikey__get",
        "parameters": [
          {
            "name": "inchikey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inchikey"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompoundView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/acquisition/run": {
      "post": {
        "tags": [
          "acquisition"
        ],
        "summary": "Acquisition Run",
        "description": "Generate hits against a user-supplied target (protein/peptide pocket).\n\nDrives the same pipeline as /acquisition/batch but with the inserted target,\nso the proposed hits respond to the input. `num_molecules` sizes the\ngeneration step; `reference_smiles` seeds MolMIM optimization when\nGENERATOR_ENGINE=molmim. The result is cached for approve.",
        "operationId": "acquisition_run_acquisition_run_post",
        "parameters": [
          {
            "name": "num_molecules",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "title": "Num Molecules"
            }
          },
          {
            "name": "reference_smiles",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference Smiles"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Target"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcquisitionBatch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/structure/fold": {
      "post": {
        "tags": [
          "structure"
        ],
        "summary": "Structure Fold",
        "operationId": "structure_fold_structure_fold_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Target"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoldResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dock": {
      "post": {
        "tags": [
          "structure"
        ],
        "summary": "Dock",
        "operationId": "dock_dock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/structure/pose": {
      "post": {
        "tags": [
          "structure"
        ],
        "summary": "Structure Pose",
        "operationId": "structure_pose_structure_pose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoseResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcquisitionBatch": {
        "properties": {
          "target_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Name"
          },
          "generated": {
            "type": "integer",
            "title": "Generated",
            "default": 0
          },
          "scored": {
            "type": "integer",
            "title": "Scored",
            "default": 0
          },
          "shortlisted": {
            "type": "integer",
            "title": "Shortlisted",
            "default": 0
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/AcquisitionCandidate"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "title": "AcquisitionBatch"
      },
...
      "AssayRecord": {
        "properties": {
          "compound": {
            "$ref": "#/components/schemas/Compound"
          },
          "assay": {
            "$ref": "#/components/schemas/Assay"
          },
          "conditions": {
            "$ref": "#/components/schemas/Conditions"
          },
          "measurement": {
            "$ref": "#/components/schemas/Measurement"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "record_id": {
            "type": "string",
            "title": "Record Id"
          },
          "status": {
            "$ref": "#/components/schemas/RecordStatus"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail",
            "description": "Human/machine-readable explanation, quoted verbatim in UI"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields",
            "description": "Exact fields needed to unblock"
          },
          "value_M": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value M",
            "description": "Reported value normalized to molar"
          },
          "ki_M": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ki M",
            "description": "Derived Ki in molar (Cheng-Prusoff)"
          },
          "comparability_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparability Key",
            "description": "Records with the same key are poolable"
          }
        },
        "type": "object",
        "required": [
          "compound",
          "assay",
          "record_id",
          "status"
        ],
        "title": "AssayRecord",
        "description": "Stored assay record: raw record + computed status + derived fields."
      },
      "AssayRecordIn": {
        "properties": {
          "compound": {
            "$ref": "#/components/schemas/Compound"
          },
          "assay": {
            "$ref": "#/components/schemas/Assay"
          },
          "conditions": {
            "$ref": "#/components/schemas/Conditions"
          },
          "measurement": {
            "$ref": "#/components/schemas/Measurement"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "type": "object",
        "required": [
          "compound",
          "assay"
        ],
        "title": "AssayRecordIn",
        "description": "Raw assay record submitted for ingest (spec \u00a71)."
      },
...
      "Compound": {
        "properties": {
          "smiles": {
            "type": "string",
            "title": "Smiles",
            "description": "Working representation (required)"
          },
          "smiles_canonical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smiles Canonical",
            "description": "Derived: RDKit canonicalization"
          },
          "inchikey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inchikey",
            "description": "Derived: stable join/dedup key"
          },
          "compound_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound Id",
            "description": "Assigned on register"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id",
            "description": "Optional salt/batch metadata"
          }
        },
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "Compound"
      },
...
      "CompoundView": {
        "properties": {
          "inchikey": {
            "type": "string",
            "title": "Inchikey"
          },
          "smiles_canonical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smiles Canonical"
          },
          "compound_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound Id"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ComparabilityGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "prediction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Prediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "measured_ki_M": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Ki M",
            "description": "Best/representative measured Ki for predicted-vs-measured delta"
          },
          "delta_loguM": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Logum",
            "description": "Boltz prediction minus measured (log-\u00b5M); calibration signal"
          }
        },
        "type": "object",
        "required": [
          "inchikey"
        ],
        "title": "CompoundView"
      },
...
      "DockRequest": {
        "properties": {
          "target": {
            "$ref": "#/components/schemas/Target"
          },
          "smiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Smiles"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "DockRequest"
      },
      "DockResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/DockResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "title": "DockResponse"
      },
...
      "FoldResult": {
        "properties": {
          "plddt": {
            "type": "number",
            "title": "Plddt",
            "description": "Mean predicted structure confidence (pLDDT)"
          },
          "pocket_residues": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pocket Residues"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "alphafold2"
          }
        },
        "type": "object",
        "required": [
          "plddt"
        ],
        "title": "FoldResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/AssayRecord"
          }
        },
        "type": "object",
        "required": [
          "record"
        ],
        "title": "IngestResponse"
      },
...
      "PoseRequest": {
        "properties": {
          "smiles": {
            "type": "string",
            "title": "Smiles"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Target"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "PoseRequest"
      },
      "PoseResult": {
        "properties": {
          "smiles": {
            "type": "string",
            "title": "Smiles"
          },
          "sdf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sdf",
            "description": "3-D docked pose as an SDF/MOL block"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "diffdock"
          }
        },
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "PoseResult"
      },
...
      "RecordPatch": {
        "properties": {
          "compound": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Compound"
              },
              {
                "type": "null"
              }
            ]
          },
          "assay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Assay"
              },
              {
                "type": "null"
              }
            ]
          },
          "conditions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Conditions"
              },
              {
                "type": "null"
              }
            ]
          },
          "measurement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Measurement"
              },
              {
                "type": "null"
              }
            ]
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RecordPatch",
        "description": "Partial edit to a record's metadata. Any sub-object provided is merged."
      },
      "RecordStatus": {
        "type": "string",
        "enum": [
          "model_ready",
          "normalizable",
          "blocked"
        ],
        "title": "RecordStatus"
      },
      "ReplayRequest": {
        "properties": {
          "inchikeys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inchikeys",
            "description": "Subset to replay; default = current acquisition batch"
          }
        },
        "type": "object",
        "title": "ReplayRequest"
      },
      "ReplayResponse": {
        "properties": {
          "replayed": {
            "type": "integer",
            "title": "Replayed"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/AssayRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "calibration_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Error"
          }
        },
        "type": "object",
        "required": [
          "replayed"
        ],
        "title": "ReplayResponse"
      },
...
      "Target": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "protein_sequence": {
            "type": "string",
            "title": "Protein Sequence"
          },
          "chain_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chain Ids"
          },
          "pocket_residues": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pocket Residues"
          }
        },
        "type": "object",
        "required": [
          "protein_sequence"
        ],
        "title": "Target"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}